+++ R/predict_mid.R
@@ -49,55 +49,59 @@ predict.mid <- function(
     terms <- unique(terms[!is.na(terms)])
   }
   if (is.null(newdata)) {
-    preds <- object$fitted.matrix[, terms, drop = FALSE]
-    naa <- stats::na.action(object)
-  } else {
-    newdata <- model.reframe(object, newdata)
-    attr(newdata, "na.action") <- NULL
-    newdata <- do.call(na.action, list(newdata))
-    naa <- stats::na.action(newdata)
-    n <- nrow(newdata)
-    r <- length(terms)
-    preds <- matrix(0, nrow = n, ncol = r)
-    colnames(preds) <- terms
-    spl <- sapply(strsplit(terms, ":"), length)
-    mts <- unique(terms[spl == 1L])
-    its <- unique(terms[spl == 2L])
-    mmats <- list()
-    for (tag in mts)
-      mmats[[tag]] <- object$encoders$main.effects[[tag]]$encode(newdata[, tag])
-    imats <- list()
-    for (tag in unique(term.split(its)))
-      imats[[tag]] <- object$encoders$interactions[[tag]]$encode(newdata[, tag])
-    for (i in seq_len(r)) {
-      term <- terms[i]
-      tags <- term.split(term)
-      if (length(tags) == 1L) {
-        X <- mmats[[term]]
-        mid <- object$main.effects[[term]]$mid
-      } else if (length(tags) == 2L) {
-        n1 <- object$encoders$interactions[[tags[1L]]]$n
-        n2 <- object$encoders$interactions[[tags[2L]]]$n
-        uni <- as.matrix(expand.grid(1L:n1, 1L:n2))
-        X <- matrix(0, nrow = n, ncol = nrow(uni))
-        for (j in seq_len(nrow(uni))) {
-          X[, j] <- as.numeric(imats[[tags[1L]]][, uni[j, 1L]] *
-                               imats[[tags[2L]]][, uni[j, 2L]])
-        }
-        mid <- object$interactions[[term]]$mid
-      }
-      preds[, i] <- as.numeric(X %*% mid)
+    newdata <- model.data(object, env = parent.frame())
+    if (is.null(newdata)) {
+      stop("'newdata' must be provided")
     }
   }
+  newdata <- model.reframe(object, newdata)
+  attr(newdata, "na.action") <- NULL
+  newdata <- do.call(na.action, list(newdata))
+  naa <- stats::na.action(newdata)
+  n <- nrow(newdata)
+  r <- length(terms)
   if (type == "terms") {
-    if (inherits(naa, "exclude"))
-      return(stats::napredict(naa, preds))
-    return(structure(preds, constant = object$intercept, na.action = naa))
+    preds <- matrix(0, nrow = n, ncol = r, dimnames = list(NULL, terms))
+  } else {
+    preds <- rep(object$intercept, n)
   }
-  preds <- rowSums(preds) + object$intercept
-  if (type == "response" && !is.null(object$link))
+  spl <- strsplit(terms, ":")
+  spllen <- sapply(spl, length)
+  mtags <- unique(terms[spllen == 1L])
+  itags <- unique(unlist(spl[spllen == 2L]))
+  mmats <- list()
+  for (tag in mtags)
+    mmats[[tag]] <- object$encoders$main.effects[[tag]]$encode(newdata[, tag])
+  imats <- list()
+  for (tag in itags)
+    imats[[tag]] <- object$encoders$interactions[[tag]]$encode(newdata[, tag])
+  for (i in seq_len(r)) {
+    tags <- spl[[i]]
+    if (length(tags) == 1L) {
+      term_preds <- as.numeric(
+        mmats[[tags]] %*% object$main.effects[[tags]]$mid
+      )
+    } else if (length(tags) == 2L) {
+      n1 <- object$encoders$interactions[[tags[1L]]]$n
+      n2 <- object$encoders$interactions[[tags[2L]]]$n
+      A1 <- imats[[tags[1L]]]
+      A2 <- imats[[tags[2L]]]
+      W <- matrix(object$interactions[[terms[i]]]$mid, nrow = n1, ncol = n2)
+      term_preds <- rowSums((A1 %*% W) * A2)
+    }
+    if (type == "terms") {
+      preds[, i] <- term_preds
+    } else {
+      preds <- preds + term_preds
+    }
+  }
+  if (type == "response" && !is.null(object$link)) {
     preds <- object$link$linkinv(preds)
-  if (inherits(naa, "exclude"))
-    return(stats::napredict(naa, preds))
-  structure(as.numeric(preds), na.action = naa)
+  } else if (type == "terms") {
+    attr(preds, "constant") <- object$intercept
+  }
+  if (inherits(naa, "exclude")) {
+    preds <- stats::napredict(naa, preds)
+  }
+  return(structure(preds, na.action = naa))
 }